core::msg::parse: add Parse trait and parse_byte_string() method
While mapping the error messages from the core/parse-sexp conformance
test to ParseErrorKind values, it became apparent that a new ExpectedEOF
kind is necessary, and that a dedicated method, parse_byte_string(), was
necessary to generate it. Since parse_byte_string() is a trivial
extension of the parse() methods found on multiple types in this module,
it feels natural to put both methods in a trait.