add vt6::server::{Connection, Handler} and vt6::core::server
This represents a development effort that has been going on for some
time, but has now been squashed and rebased onto master to account for
the recent changes to vt6::core::msg.

The current implementation of handle_want() is not going to fly: The
message-formatting closure for the positive `have` answer should outlive
the `module_name` reference that it takes it. I believe that the borrow
checker does not complain because the signature of
`core::server::Handler::handle()` requires the same lifetime for
`&Message` and `&mut Connection`, which is not possible since `&Message`
has the same lifetime as the receive buffer in the connection, but we
cannot use `&mut Connection` while also using the immutable reference to
the receive buffer that lives in the connection.