make server::Connection actually implementable
The problem with the version in the previous commit was that
the closures given to `write_to_send_buffer()` would have needed to be
'static if they outlive the `write_to_send_buffer()` call. Both
situations are not plausible in a no_std scenario if it is possible that
the remaining free space in the send buffer is too small to render
responses to.

Therefore, change the contract on Connection::write_to_send_buffer (now
called Connection::with_send_buffer) to require eager evaluation of the
action, *but* to make this work, also change Handler::handle to be
restartable in case of errors. The error type returned from Handler
indicates how much additional space in the send buffer would be required
to successfully answer the request.

To make this error type work, also change MessageFormatter to track the
number of written bytes beyond the point of overflow. Since
BufferTooSmallError is now only returned from finalize(), format() is no
longer required to keep the code clean, and therefore is removed.