vt6::core::server::Handler: implement core.set and core.sub handling
This requires yet another change to how send buffers are passed into the
handler: get_set_property() returns an Option<&EncodeArgument> whose
lifetime is bound to a borrow of Connection. Therefore we cannot borrow
Connection again (mutably!) to call Connection::with_send_buffer().

To solve this, the send buffer is now passed into handle() separately,
so that the connection state and the send buffer can be used at the same
time.