replace Handler::get_set_property() by Handler::handle_property()
The previous design turned out not to work in practice. When returning
an `Option<&'c EncodeArgument>`, the underlying type of trait object
`&EncodeArgument` needs to be `Sized`, so we could not return `&str` or
`&[u8]`.

The new design works for string-typed properties, as evidenced by the
implementation of `TestHandler` in src/core/server/tests.rs. It remains
to be discussed if this design is sufficiently nice to stay.