start to build the server implementation
Not pictured: Over a week worth of back and forth on the overall
structure, trying to come to an agreement with rustc on whether my
lifetimes and ownerships make sense.

The current setup is missing a few crucial pieces:
- support for stdin and stdout mode in handle_incoming
- the initial set of Handler objects for handshaking and core
- the enqueue_broadcast implementation

But I'm finally somewhat confident that what is currently there can
accomodate what's missing without that much change.