add a lifetime arg to trait DecodeArgument
I wasn't aware that one can do this until now. With this change, all the
identifier types in vt6::common:core can impl this trait via their
existing parse() methods.

I also aligned this trait with DecodeMessage by changing the method name
from decode() to decode_argument(). This is useful because the decoder
traits are typically pulled in on the module scope, so name collisions
would be bad for ergonomics. The encoder traits stay with the short
method names. Those are not ambiguous since those methods are only
called in a small number of very specific places.