move {client,parent,server}-hello messages from vt6/foundation to vt6/posix1
These messages are too platform-dependent to live in the foundation
spec. I initially put them in foundation because they have to be
eternal, but then again, there's a lot of de-facto eternal stuff in
vt6/posix1 already, so we don't make anything worse by putting the
messages where they belong.